add date check + add at logic

+++ delete.php
@@ -1,10 +1,18 @@
 <?php
-
     include_once 'config.php';
 
-    $jsonDb->delete()
-        ->from( DB_TABLE_JOB )
-        ->where( [ 'id' => $_GET['id'] ] )
-        ->trigger();
+    $job = $jsonDb->select('*')->from(DB_TABLE_JOB)->where(['id' => $_GET['id']])->getOne();
+
+    if (!empty($job)) {
+        if (!empty($job['atId'])) {
+            exec('atrm ' . $job['atId']);
+        }
+
+        $jsonDb->delete()
+            ->from( DB_TABLE_JOB )
+            ->where( [ 'id' => $job['id'] ] )
+            ->trigger();
+    }
 
     Header('Location:index.php');
+?>
